CoreNetworking:
Added a warning to use the service values in /etc/services using the functions in netdb.h
In the AFNetworkTransport, it now only sets the close sockets property on the streams if it was initially set on the socket. This allows it to use sockets that it shouldn't take ownership of.