AmberFoundation:
Trying to fix an old bug here. When breaking in the AFPropertyList the debugger tries to get the count of self, seeing that it's an array. This causes Cocoa to raise an exception because the array is uninitialised. By autoreleasing it and setting to nil immediately the likelyhood of this happening is reduced - but not removed.

If you see an exception raised in these methods, that it's unlikely to be an error on your part.