Core Networking:
Added @private to the instance variables as appropriate.
Rewrote the packet queuing in `AFNetworkTransport`, bringing the need to separate the queuing into another class to the forefront again.
Added two new packet types, `AFPacketWriteFromReadStream` and `AFPacketReadToWriteStream`.
Changed the packet API to conform to the NSError by-ref returning API contract properly.